+++ lib/DBSQLOperation/index.ts
@@ -1,5 +1,5 @@
 import { stringify, NIL, parse } from 'uuid';
-import IOperation, { FetchOptions, GetSchemaOptions, FinishedOptions, defaultMaxRows } from '../contracts/IOperation';
+import IOperation, { FetchOptions, GetSchemaOptions, FinishedOptions } from '../contracts/IOperation';
 import HiveDriver from '../hive/HiveDriver';
 import {
   TGetOperationStatusResp,
@@ -15,6 +15,8 @@ import SchemaHelper from './SchemaHelper';
 import FetchResultsHelper from './FetchResultsHelper';
 import CompleteOperationHelper from './CompleteOperationHelper';
 import IDBSQLLogger, { LogLevel } from '../contracts/IDBSQLLogger';
+
+const defaultMaxRows = 100000;
 
 export default class DBSQLOperation implements IOperation {
   private driver: HiveDriver;